+++ tests/testthat/test-get_varcov.R
@@ -82,3 +82,19 @@ test_that("error: ill-defined model", {
   m1 <- lm(y ~ 1, data = dd)
   expect_error(get_varcov(m1), regex = "Can't extract variance-covariance")
 })
+
+
+test_that("error: bad string", {
+  m <- lm(mpg ~ hp + factor(carb), data = mtcars)
+  expect_error(get_varcov(m, vcov = "bootstrap"), regexp = "not a recognized")
+})
+
+
+test_that("error: from vcov", {
+  data(iris)
+  mod <- lm(Sepal.Width ~ Sepal.Length, data = iris)
+  expect_error(
+    get_varcov(mod, vcov = "vcovBS", vcov_args = list(cluster = "Species")),
+    regex = "number of observations"
+  )
+})